Block and Command ItemIds are now ushorts | Var Ids are bytes | Refactored EventHandler subclasses for cleanliness | EventHandler base class more convenient
Had to refactor various other parts of the codebase, what with the id type changes

+++ Assets/Amanita/Systems/NewSaveSys/GameStateContents/Codecs/BlockSaveCodec.cs
@@ -53,7 +53,7 @@ namespace Amanita.SaveSys
         public override BlockSaveData EncodeToSave(Block toCreateFrom)
         {
             // We assume that the Block was indeed executing at this point.
-            int itemId = toCreateFrom.ItemId;
+            ushort itemId = toCreateFrom.ItemId;
             string blockName = toCreateFrom.BlockName;
             int activeCommandId = -1, activeCommandIndex = -1;
